Rename recursion -> pizza-order
