+++ exercises/concept/pizza-order/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@exercism/javascript-recursion",
+  "name": "@exercism/javascript-pizza-order",
   "description": "Exercism concept exercise on recursion",
   "author": "Derk-Jan Karrenbeld <derk-jan+github@karrenbeld.info>",
   "private": true,
@@ -7,7 +7,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/exercism/javascript",
-    "directory": "exercises/concept/recursion"
+    "directory": "exercises/concept/pizza-order"
   },
   "devDependencies": {
     "@babel/cli": "^7.14.5",
